requirements.txt
h5py==2.8.0
imageio==2.5.0
imagesize==0.7.1
matplotlib==3.0.3
numpy==1.16.3
opencv-python==3.4.3.18
pandas==0.18.1
Pillow==5.2.0
PyYAML==3.13
scikit-image==0.15.0
scikit-learn==0.19.1
scikit-surprise==1.0.4
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorboardX==1.4
torch==0.4.0
torchvision==0.2.1
plyfile==0.6
